extracts the population density at the locality level in hexagons of 400 meters resolution for the years 2023 and 2022

extracts the population density at the locality level in hexagons of 400 meters resolution for the years 2023 and 2022

extract_pop(year, cve_edo, locality)

Arguments

year

is the year dataset

cve_edo

is the id of state.

locality

is the target locality.

Value

a sf object

Author

Felipe Antonio Dzul Manzanilla felipe.dzul.m@gmail.com.

Examples

popmex::extract_pop(year = 2023, cve_edo = "31", locality = "Merida")
#> Simple feature collection with 401 features and 2 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -89.73667 ymin: 20.8754 xmax: -89.51838 ymax: 21.09618
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>                     h3 population                           geom
#> 427027 8845724d27fffff          5 POLYGON ((-89.71474 20.9272...
#> 427028 8845724d25fffff        935 POLYGON ((-89.71032 20.9198...
#> 427172 88457249bbfffff       1299 POLYGON ((-89.65264 20.8833...
#> 427173 88457249b9fffff        135 POLYGON ((-89.64824 20.8758...
#> 427174 88457249b7fffff       1652 POLYGON ((-89.63924 20.8910...
#> 427175 88457249b5fffff         31 POLYGON ((-89.63483 20.8836...
#> 427176 88457249b3fffff       1401 POLYGON ((-89.64814 20.8908...
#> 427177 88457249b1fffff        127 POLYGON ((-89.64374 20.8834...
#> 427181 88457249a7fffff        190 POLYGON ((-89.62153 20.8763...
#> 427183 88457249a3fffff        240 POLYGON ((-89.63043 20.8761...